requirements.txt
langchain
openai
python-dotenv
beautifulsoup4
llama_index
